updated error in seed data

+++ db/schema.sql
@@ -8,6 +8,6 @@ CREATE TABLE users (
     country VARCHAR(255),
     photo VARCHAR(255),
     about TEXT,
-    latitude DECIMAL(9, 6),
-    longitude DECIMAL(9, 6)
+    latitude FLOAT(9, 6),
+    longitude FLOAT(9, 6)
 );
